refactor: pulled out processing & spec name finding into helpers
